requirements.txt
PyJWT==2.10.1
sendgrid==6.11.0

# Core Telegram Bot
python-telegram-bot==22.0

# Firebase Database
firebase-admin==6.7.0
google-cloud-firestore==2.20.1
google-auth==2.38.0

# PDF Generation
fpdf==1.7.2

# HTTP/API Clients
requests==2.32.3
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
uuid==1.30
pyparsing==3.2.1

# Security
cryptography==44.0.2

